requirements.txt
defusedxml==0.4.1
Django==1.7.7
django-allauth==0.19.1
django-axes==1.3.9
django-debug-toolbar==1.3.0
django-rest-auth==0.3.4
djangorestframework==3.1.1
oauthlib==0.7.2
python3-openid==3.0.5
requests==2.6.0
requests-oauthlib==0.4.2
six==1.9.0
sqlparse==0.1.14
